+++ Sources/NIOSOCKS/Messages/Errors.swift
@@ -78,4 +78,7 @@ public enum SOCKSError {
         }
     }
     
+    struct MissingBytes: Error {
+        
+    }
 }
